Setup Homepage HTML and Create Learning Pages
Created the index.html file and filled it with some description on what
the website if for. This includes the learning courses provided, about
the creator, and links to the courses.
Created the images and pages directories to hold the images and seperate
html files the website will use.
Inside the pages directory, the HTML pages for the programming courses
have been initialized without any content.